[gokillctl] mv subcommands into separate files
